datapath: Add a new action dec_ttl
Add support for the dec_ttl action. Instead of programming the datapath with
a flow that matches the packet TTL and an IP set, use a single dec_ttl action.

The old behavior is kept if the new action is not supported by the datapath.

  # ovs-ofctl dump-flows br0
   cookie=0x0, duration=12.538s, table=0, n_packets=4, n_bytes=392, ip actions=dec_ttl,NORMAL
   cookie=0x0, duration=12.536s, table=0, n_packets=4, n_bytes=168, actions=NORMAL

  # ping -c1 -t 20 192.168.0.2
  PING 192.168.0.2 (192.168.0.2) 56(84) bytes of data.
  IP (tos 0x0, ttl 19, id 45336, offset 0, flags [DF], proto ICMP (1), length 84)
      192.168.0.1 > 192.168.0.2: ICMP echo request, id 8865, seq 1, length 64

Linux netlink datapath support depends on upstream Linux commit:
  744676e77720 ("openvswitch: add TTL decrement action")


Note that in the Linux kernel tree the OVS_ACTION_ATTR_ADD_MPLS has been
defined, and to make sure the IDs are in sync, it had to be added to the
OVS source tree. This required some additional case statements, which
should be revisited once the OVS implementation is added.


Co-developed-by: Matteo Croce <mcroce@linux.microsoft.com>
Co-developed-by: Bindiya Kurle <bindiyakurle@gmail.com>
Signed-off-by: Eelco Chaudron <echaudro@redhat.com>


v2: Used definition instead of numeric value in format_dec_ttl_action()
    Changed format from "dec_ttl(ttl<=1(<actions>)) to "dec_ttl(le_1(<actions>))" to be more in line with the check_pkt_len action.
    Fixed parsing of "dec_ttl()" action for adding a dp flow.