utilities: Add revalidator measurement script and needed USDT probes.
This patch adds a Python script that can be used to analyze the
revalidator runs by providing statistics (including some real time
graphs).

The USDT events can also be captured to a file and used for
later offline analysis.

Signed-off-by: Eelco Chaudron <echaudro@redhat.com>

v2: Added note that script only works a with single datapath configured.

I'm planning a blog post to explain the Open vSwitch revalidator
implementation, and this tool will help you determine what is
happening in your system. Also, here is a link to an example
of a real-time and overall plot.

  https://photos.app.goo.gl/rdx63zuFure7QE3t6