netdev-offload-tc: Fix ignore of dl type mask when installing flows.
When TC parses the flow to install, it assumes that the datalink type
mask is set. However, this may not always be the case, for example,
when multiple VLANs exist but only one is enabled (vlan-limit).

This patch will only process the dl_type if the mask is set. It also
includes a unit test to verify that the TC rules are offloaded in this
case.

Fixes: 1be33d52af77 ("netdev-tc-offloads: Don't offload header modification on ip fragments.")
Signed-off-by: Eelco Chaudron <echaudro@redhat.com>

---
v3: - Fixed all Ilya's style comments
v2: - Adding is_ip_any() function that got deleted somehow
    - Fixed alignment