tests: Fix multiple Coverity warnings in test programs.
This patch addresses several common issues in various test programs
instead of sending multiple small patches.  While these are not
critical, cleaning them up improves code quality.

- Fixed two resource leaks in test-aa.c and test-ovsdb.c.
- Fixed four instances of unchecked return values in test-ovsdb.c
  and test-packet.c.
- Fixed one out-of-bounds read in test-ccmap.c

Signed-off-by: Eelco Chaudron <echaudro@redhat.com>
---
Note that 'ERROR: Use ovs_assert() in place of assert()' is
reported. However we kept assert() as it's the standard for
this file.