tests: Add a simple DPDK rte_flow test framework.
This patch adds userspace DPDK with rte_flow offload unit
tests. It can be run with the 'make check-dpdk-offloads'
command; however, it requires a list of VFs that can be
used for offloaded ports. More details on how to set this
up can be found in the testing.rst file.

Also note that currently not all tests are passing. This is
true for both the old and the new offload infrastructure:

 - datapath - ping over ip6gre L2 tunnel
   This test appears to be failing because the hardware offload
   NIC is not properly handling checksum offloads. Specifically,
   the L4 checksum is not updated, and as a result the kernel
   drops the packets.

 - mpls - decap header dp-support
 - mpls - decap header slow-path
   These two tests fail due to side traffic: an IPv6 drop
   rule gets installed in hardware. However, this rule also
   drops MPLS-encapsulated IPv6 traffic. This appears to be
   a bug specific to the hardware offload NIC in use.

Signed-off-by: Eelco Chaudron <echaudro@redhat.com>
---

v2 changes:
  - Fixed some typos.