dpif-offload: Add infrastructure for offload provider PMD helpers.
This patch adds support for specific PMD thread initialization,
deinitialization, and a callback execution to perform work as
part of the PMD thread loop. This allows hardware offload
providers to handle any specific asynchronous or batching work.

This patch also adds cycle statistics for the provider-specific
callbacks to the 'ovs-appctl dpif-netdev/pmd-perf-show' command.

Signed-off-by: Eelco Chaudron <echaudro@redhat.com>
---

Note that this patch builds on top of the hardware offload rework
patch set, which can be found here:
https://patchwork.ozlabs.org/project/openvswitch/list/?series=484144

v2:
  - Added offload cycle statistics to the 'ovs-appctl
    dpif-netdev/pmd-perf-show' command.